test(coding-agent): relocate issue regressions (fixes #931)
