+++ packages/coding-agent/test/suite/regressions/931-package-pattern-dependency.test.ts
@@ -2,8 +2,8 @@ import { mkdirSync, rmSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { DefaultPackageManager } from "../src/core/package-manager.js";
-import { SettingsManager } from "../src/core/settings-manager.js";
+import { DefaultPackageManager } from "../../../src/core/package-manager.js";
+import { SettingsManager } from "../../../src/core/settings-manager.js";
 
 describe("package pattern dependency", () => {
 	let tempDir: string;
